fixing big integers again

+++ tla-bmcmt/src/main/scala/at/forsyte/apalache/tla/bmcmt/rules/IntConstRule.scala
@@ -21,7 +21,7 @@ class IntConstRule(rewriter: SymbStateRewriter) extends RewritingRule {
   override def apply(state: SymbState): SymbState = {
     state.ex match {
       case ValEx(TlaInt(n)) =>
-        val (newArena: Arena, intCell: ArenaCell) = rewriter.intValueCache.getOrCreate(state.arena, n.toInt)
+        val (newArena: Arena, intCell: ArenaCell) = rewriter.intValueCache.getOrCreate(state.arena, n)
         state.setArena(newArena)
           .setRex(intCell.toNameEx)
